% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossAPI_function.R
\name{API argument. Defaults to NA}
\alias{API argument. Defaults to NA}
\alias{fossAPI}
\title{API Wrapper for NOAA FOSS Database}
\source{
API argument. Defaults to NA

_species_code API argument. Defaults to NA

_id API argument. Defaults to NA

_gear_code API argument. Defaults to NA
}
\usage{
fossAPI(
  series = NA,
  allow_duplicates = FALSE,
  list_variables = FALSE,
  tsn = NA,
  ts_afs_name = NA,
  in_ts_afs_name = NA,
  ts_scientific_name = NA,
  in_ts_scientific_name = NA,
  region_name = NA,
  state_name = NA,
  year = NA,
  month = NA,
  pounds = NA,
  dollars = NA,
  tot_count = NA,
  source = NA,
  collection = NA,
  hts_number = NA,
  name = NA,
  in_name = NA,
  cntry_code = NA,
  fao = NA,
  cntry_name = NA,
  in_cntry_name = NA,
  district_code = NA,
  district_name = NA,
  in_district_name = NA,
  edible_code = NA,
  kilos = NA,
  val = NA,
  association = NA,
  rfmo = NA,
  nmfs_region_code = NA,
  source_species_code = NA,
  ss_name = NA,
  in_ss_name = NA,
  nmfs_species_code = NA,
  source_id = NA,
  tsn_species_id = NA,
  gear_id = NA,
  isscfg = NA,
  g_name_short = NA,
  g_name_long = NA,
  in_g_name_short = NA,
  in_g_name_long = NA,
  g_desc = NA,
  g_level_ind = NA,
  g_parent_id = NA,
  old_itis = NA,
  new_itis = NA,
  source_gear_code = NA,
  sg_name = NA,
  in_sg_name = NA,
  ts_complex = NA,
  ts_confidential_name = NA,
  in_ts_confidential_name = NA,
  ts_finfish_ind = NA,
  ts_habitat_ind = NA,
  protected = NA,
  reported = NA,
  family_itis = NA,
  family_name = NA,
  in_family_name = NA,
  superclass_itis = NA,
  superclass_name = NA,
  in_superclass_name = NA,
  ts_kingdom = NA,
  in_ts_kingdom = NA,
  genus_itis = NA,
  genus_name = NA,
  in_genus_name = NA,
  family_sc_name = NA,
  in_family_sc_name = NA,
  genus_sc_name = NA,
  in_genus_sc_name = NA
)
}
\description{
Use simple notation to query the APIs in the FOSS database.
}
\concept{_itis API argument. Defaults to NA}
\concept{_name API argument. Defaults to NA}
\concept{_sc_name API argument. Defaults to NA}
